+++ src/event/base.js
@@ -39,8 +39,16 @@ KISSY.add('event/base', function(S, DOM, EventObject, undefined) {
 
 
     var Event = {
-
-        EVENT_GUID: EVENT_GUID,
+        _data:function(elem) {
+            var args = S.makeArray(arguments);
+            args.splice(1, 0, EVENT_GUID);
+            return DOM.data.apply(DOM, args);
+        },
+        _removeData:function(elem) {
+            var args = S.makeArray(arguments);
+            args.splice(1, 0, EVENT_GUID);
+            return DOM.removeData.apply(DOM, args);
+        },
 
         // such as: { 'mouseenter' : { setup:fn ,tearDown:fn} }
         special: { },
@@ -76,9 +84,9 @@ KISSY.add('event/base', function(S, DOM, EventObject, undefined) {
 
 
                 // 获取事件描述
-                eventDesc = DOM.data(target, EVENT_GUID);
+                eventDesc = Event._data(target);
                 if (!eventDesc) {
-                    DOM.data(target, EVENT_GUID, eventDesc = {});
+                    Event._data(target, eventDesc = {});
                 }
                 //事件 listeners
                 events = eventDesc.events = eventDesc.events || {};
@@ -133,7 +141,7 @@ KISSY.add('event/base', function(S, DOM, EventObject, undefined) {
 
         __getEvents:function(target) {
             // 获取事件描述
-            var eventDesc = DOM.data(target, EVENT_GUID);
+            var eventDesc = Event._data(target);
             return eventDesc && eventDesc.events;
         },
 
@@ -146,7 +154,7 @@ KISSY.add('event/base', function(S, DOM, EventObject, undefined) {
             }
 
             DOM.query(targets).each(function(target) {
-                var eventDesc = DOM.data(target, EVENT_GUID),
+                var eventDesc = Event._data(target),
                     events = eventDesc && eventDesc.events,
                     listeners,
                     len,
@@ -220,7 +228,7 @@ KISSY.add('event/base', function(S, DOM, EventObject, undefined) {
                     eventDesc.handler.target = null;
                     delete eventDesc.handler;
                     delete eventDesc.events;
-                    DOM.removeData(target, EVENT_GUID);
+                    Event._removeData(target);
                 }
             });
             return targets;
@@ -276,7 +284,7 @@ KISSY.add('event/base', function(S, DOM, EventObject, undefined) {
                 eventData = eventData || {};
                 eventData.type = eventType;
                 if (!isNativeEventTarget) {
-                    var eventDesc = DOM.data(target, EVENT_GUID);
+                    var eventDesc = Event._data(target);
                     if (eventDesc && S.isFunction(eventDesc.handler)) {
                         ret = eventDesc.handler(undefined, eventData);
                     }
@@ -290,7 +298,7 @@ KISSY.add('event/base', function(S, DOM, EventObject, undefined) {
                         ontype = "on" + eventType;
                     //bubble up dom tree
                     do{
-                        var handler = (DOM.data(cur, EVENT_GUID) || {}).handler;
+                        var handler = (Event._data(cur) || {}).handler;
                         event.currentTarget = cur;
                         if (handler) {
                             handler.call(cur, event);
@@ -365,7 +373,7 @@ KISSY.add('event/base', function(S, DOM, EventObject, undefined) {
     }
 
     if (1 > 2) {
-        Event._simpleAdd()._simpleRemove(Event.EVENT_GUID);
+        Event._simpleAdd()._simpleRemove();
     }
 
     return Event;