+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Legend of version history:
 History:
 --------
 ### v1.1.6 -> v1.2.0 (2011/06/08)
+ - [+] event 增加作用于 dom 节点的 delegate 方法
+ - [+] event 增加作用于 dom 节点的 fire 方法
  - [*] 自定义事件 listeners 放入对象自身保存，避免内存泄露
  - [+] event 增加 valuechange ,hashchange 事件兼容处理
  - [*] Event.detach = Event.remove